fetch contribution on filters update
