+++ web/src/apps/main/redux/reducers/contribute-page/index.ts
@@ -10,7 +10,7 @@ interface Filter {
   }>;
 }
 
-interface Contribution {
+export interface Contribution {
   id: string;
   title: string;
   description: string;
@@ -22,7 +22,7 @@ interface Contribution {
 
 export interface ContributePageState {
   filters: Filter[];
-  contributions?: Contribution[];
+  contributions: Contribution[] | null;
 }
 
 export const contributePage = (
@@ -74,142 +74,7 @@ export const contributePage = (
         ],
       },
     ],
-    contributions: [
-      {
-        id: "github/dzcode-io/dzcode.io/308",
-        title: "Adding projects and articles directly from website",
-        description: `When I clicked on the contribute button that takes to [www.dzcode.io/Contribute](https://www.dzcode.io/Contribute) I found that this page is not implemented yet and thus it rises an idea for me and that is what if there is a way to add projects and articles directly from the website without forking and adding to json files like we always do.
-
-Other than that, this website is one of the awesome ones and definitely I'm proud of it.
-
-![image](https://user-images.githubusercontent.com/48713070/119314037-26e58d00-bc6c-11eb-87b7-c35850cdef9f.png)
-`,
-        projectId: "github/dzcode-io/dzcode.io",
-        url: "https://github.com/dzcode-io/dzcode.io/issues/308",
-        languages: ["TypeScript", "JavaScript", "Pug", "SCSS", "Dockerfile"],
-        labels: ["enhancement", "ui/ux"],
-      },
-      {
-        id: "github/dzcode-io/dzcode.io/306",
-        title: "Perf: Improve initial page load experience",
-        description: `**Describe the bug**
-
-As a user/crawler, I should see minimal content when pages are loading. And ideally, the page loads faster for me.
-
-
-**To Reproduce**
-Steps to reproduce the behavior:
-
-1. Empty cache and access to any page
-
-**Expected behavior**
-
-- See "something" while the page is loading.
-- Load the main/important content at first, instead of none or everything
-
-**Screenshots**
-
-Google lighthouse score might also confirm this issue:
-
-![image](https://user-images.githubusercontent.com/9090674/119261989-ee956e80-bbd9-11eb-9582-23896d14a4eb.png)
-
-**Possible solutions**
-
-- Prioritize content loading over unnecessary assets that can be lazy-loaded (for example article content over GH contributors)
-- Dynamically render components
-- Cache Backend-to-data source requests/queries
-- Consider SSR
-- Do more work on the backend (for example fetching contributors lists and returning their avatars).`,
-        projectId: "github/dzcode-io/dzcode.io",
-        url: "https://github.com/dzcode-io/dzcode.io/issues/306",
-        languages: ["TypeScript", "JavaScript", "Pug", "SCSS", "Dockerfile"],
-        labels: ["bug", "good first issue"],
-      },
-      {
-        id: "github/dzcode-io/kuliya/13",
-        title: "add Java support",
-        description: `**Is your feature request related to a problem? Please describe.**
-we need java package for kuliya
-
-**Describe the solution you'd like**
-a package under \`./java \` folder
-
-**Resources/Links**
-- http://maven.apache.org/plugin-developers/index.html
-`,
-        projectId: "github/dzcode-io/kuliya",
-        url: "https://github.com/dzcode-io/kuliya/issues/13",
-        languages: ["TypeScript", "JavaScript", "Pug", "SCSS", "Dockerfile"],
-        labels: ["bug", "good first issue"],
-      },
-      {
-        id: "github/dzcode-io/dzcode.io/308",
-        title: "Adding projects and articles directly from website",
-        description: `When I clicked on the contribute button that takes to [www.dzcode.io/Contribute](https://www.dzcode.io/Contribute) I found that this page is not implemented yet and thus it rises an idea for me and that is what if there is a way to add projects and articles directly from the website without forking and adding to json files like we always do.
-
-Other than that, this website is one of the awesome ones and definitely I'm proud of it.
-
-![image](https://user-images.githubusercontent.com/48713070/119314037-26e58d00-bc6c-11eb-87b7-c35850cdef9f.png)
-`,
-        projectId: "github/dzcode-io/dzcode.io",
-        url: "https://github.com/dzcode-io/dzcode.io/issues/308",
-        languages: ["TypeScript", "JavaScript", "Pug", "SCSS", "Dockerfile"],
-        labels: ["enhancement", "ui/ux"],
-      },
-      {
-        id: "github/dzcode-io/dzcode.io/306",
-        title: "Perf: Improve initial page load experience",
-        description: `**Describe the bug**
-
-As a user/crawler, I should see minimal content when pages are loading. And ideally, the page loads faster for me.
-
-
-**To Reproduce**
-Steps to reproduce the behavior:
-
-1. Empty cache and access to any page
-
-**Expected behavior**
-
-- See "something" while the page is loading.
-- Load the main/important content at first, instead of none or everything
-
-**Screenshots**
-
-Google lighthouse score might also confirm this issue:
-
-![image](https://user-images.githubusercontent.com/9090674/119261989-ee956e80-bbd9-11eb-9582-23896d14a4eb.png)
-
-**Possible solutions**
-
-- Prioritize content loading over unnecessary assets that can be lazy-loaded (for example article content over GH contributors)
-- Dynamically render components
-- Cache Backend-to-data source requests/queries
-- Consider SSR
-- Do more work on the backend (for example fetching contributors lists and returning their avatars).`,
-        projectId: "github/dzcode-io/dzcode.io",
-        url: "https://github.com/dzcode-io/dzcode.io/issues/306",
-        languages: ["TypeScript", "JavaScript", "Pug", "SCSS", "Dockerfile"],
-        labels: ["bug", "good first issue"],
-      },
-      {
-        id: "github/dzcode-io/kuliya/13",
-        title: "add Java support",
-        description: `**Is your feature request related to a problem? Please describe.**
-we need java package for kuliya
-
-**Describe the solution you'd like**
-a package under \`./java \` folder
-
-**Resources/Links**
-- http://maven.apache.org/plugin-developers/index.html
-`,
-        projectId: "github/dzcode-io/kuliya",
-        url: "https://github.com/dzcode-io/kuliya/issues/13",
-        languages: ["TypeScript", "JavaScript", "Pug", "SCSS", "Dockerfile"],
-        labels: ["bug", "good first issue"],
-      },
-    ],
+    contributions: null,
   },
   action: Action<ContributePageState>,
 ) => {
